archive/projects_nested/billboard-splitter/requirements_utf8.txt
aioftp==0.25.1
altgraph==0.17.4
cffi==1.17.1
chardet==5.2.0
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
lxml==5.3.1
mysql-connector-python==9.3.0
packaging==24.2
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.2.1
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyModbusTCP==0.3.0
PyMuPDF==1.26.1
PyMySQL==1.1.1
pypdf==5.4.0
PyPDF2==3.0.1
python-barcode==0.15.1
pywin32-ctypes==0.2.3
qrcode==8.0
reportlab==4.3.1
setuptools==78.0.2
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
ttkthemes==3.2.2
types-cryptography==3.3.23.2
webencodings==0.5.1
